Refactor platform-specific functions
