+++ tetrice.c
@@ -1,7 +1,7 @@
 #include <stdint.h>
 #include <stdio.h>
 
-#include "alice.h"
+#include "platform.h"
 #include "tetromino.h"
 
 // Colors for each tetromino
@@ -15,251 +15,10 @@ char tetrominos_colors[] = {
 #define BOUNDS_X2 26
 
 /************************************************************/
-/* Macros                                                   */
+/* Platform specific functions are implemented elsewhere     */
 /************************************************************/
 
-#define POKE(addr, value) (*((volatile uint8_t *)(addr)) = (value))
-
-#define PEEK(address) (*((volatile uint8_t *)(address)))
-
-// Wait for EF9345
-#define BUSY() while (PEEK(R0) & 0x80) {}
-
-/************************************************************/
-/* Display functions for Alice 32/90                        */
-/************************************************************/
-
-// Set position registers
-void posxy(unsigned char column, unsigned char line)
-{
-    // Lines 1 to 7 do not exist... skip them
-    if (line > 0)
-        line += 7;
-
-    POKE(R6, line);   // R6 (MP / LINE)
-    POKE(R7, column); // R7 (MP / COL)
-}
-
-// Set color registers
-void color(unsigned char foreground, unsigned char background)
-{
-    unsigned char b = 0, a;
-    if (foreground > pink)
-    {
-        b = 1;
-        foreground -= 8;
-    }
-    if (background > pink)
-    {
-        b = 1;
-        background -= 8;
-    }
-    a = background + (foreground * 16);
-    POKE(R2, b); // R2 (B)
-    POKE(R3, a); // R3 (A) color
-}
-
-// Display a string with position
-void prints(unsigned char x, unsigned char y, unsigned char *text)
-{
-    char *c = text;
-    posxy(x, y);
-    for (; *c != '\0'; c++)
-    {
-        POKE(R1, *c);
-        POKE(R0EXEC, 1);
-        BUSY();
-    }
-}
-
-// Display a graphical string with position
-void printsg(unsigned char x, unsigned char y, unsigned char *text)
-{
-    char *c = text;
-    posxy(x, y);
-    for (; *c != '\0'; c++)
-    {
-        POKE(R1, *c);
-        POKE(R2, 0x20);
-        POKE(R0EXEC, 1);
-        BUSY();
-    }
-}
-
-// Display a char with position
-void printc(unsigned char x, unsigned char y, unsigned char c)
-{
-    posxy(x, y);
-    POKE(R1, c);
-    POKE(R0EXEC, 1);
-    BUSY();
-}
-
-// Display a graphical char with position
-void printcg(unsigned char x, unsigned char y, unsigned char c)
-{
-    posxy(x, y);
-    POKE(R1, c);
-    POKE(R2, 0x20);
-    POKE(R0EXEC, 1);
-    BUSY();
-}
-
-// Get character at position
-uint8_t charatxy(uint8_t column, uint8_t line)
-{
-    // Lines 1 to 7 do not exist... skip them
-    if (line > 0)
-        line += 7;
-
-    POKE(R6, line);   // R6 (MP / LINE)
-    POKE(R7, column); // R7 (MP / COL)
-
-    POKE(R0EXEC, 8);
-    BUSY();
-
-    return PEEK(R1);
-}
-
-/************************************************************/
-/* Keyboard and clock                                       */
-/************************************************************/
-
-// Scan the keyboard and return the key pressed, or 0 if no key is pressed.
-uint8_t scankey()
-{
-    uint8_t key;
-    uint8_t col;
-    uint8_t rank;
-
-    for (col = 0; col < 8; col++)
-    {
-        // Column to scan
-        POKE(0x0002, all_key_columns[col]);
-        // Get the result (row)
-        key = PEEK(0xBFFF);
-
-        // Nothing pressed, go to next column
-        if (key == 0xFF)
-            continue;
-
-        // Invert the bits
-        key = ~key;
-
-        // Find the rank of the "on" bit
-        rank = 0;
-        while ((key & 0x01) == 0)
-        {
-            key = key >> 1;
-            rank++;
-        }
-
-        // Not a valid key, go to next column
-        if (rank > 5)
-            continue;
-
-        // Return the key
-        return keys_per_column[col][rank];
-    }
-
-    // No key found
-    return 0;
-}
-
-// Sleep for a number of seconds
-void sleep(uint8_t seconds)
-{
-    uint8_t clock;
-    uint8_t tick = 0;
-
-    while (1)
-    {
-        clock = PEEK(0x0008); // clock control register
-        PEEK(0x0009);         // need to read 0x09/0x0A to reset the TOF control bit
-
-        // wait for clock bit 5 to be set
-        if (clock & 0x20)
-        {
-            tick++;
-            // 15 ticks == 1 second
-            if (tick == 15 * seconds)
-                break;
-        }
-    }
-}
-
-// Sleep for a number of ticks
-void ticks(uint8_t ticks)
-{
-    uint8_t clock;
-    uint8_t tick = 0;
-
-    while (1)
-    {
-        clock = PEEK(0x0008); // clock control register
-        PEEK(0x0009);         // need to read 0x09/0x0A to reset the TOF control bit
-
-        // wait for clock bit 5 to be set
-        if (clock & 0x20)
-        {
-            tick++;
-            if (tick == ticks)
-                break;
-        }
-    }
-}
-
-// Wait for key or timeout in ticks
-// Remember: 15 ticks == 1 second
-
-// Global variable to store the timeout in ticks
-// this needs to be set before calling wait()
 uint8_t timeout_ticks = 0;
-
-uint8_t wait()
-{
-    uint8_t clock;
-    uint8_t tick = 0;
-    uint8_t c;
-
-    while (1)
-    {
-        clock = PEEK(0x0008); // clock control register
-        PEEK(0x0009);         // need to read 0x09 to reset the TOF control bit
-
-        // wait for clock bit 5 to be set
-        if (clock & 0x20)
-        {
-            tick++;
-            // if timeout is reached, return 0
-            if (tick >= timeout_ticks)
-                return 0;
-        }
-
-        // scan the keyboard
-        c = scankey();
-        if (c != 0)
-        {
-            // if a key is pressed, decrement the timeout
-            timeout_ticks -= tick;
-            // and return the key
-            return c;
-        }
-    }
-}
-
-// Wait for key forever
-uint8_t wait_key()
-{
-    uint8_t c;
-
-    while (1)
-    {
-        c = scankey();
-        if (c != 0)
-            return c;
-    }
-}
 
 /************************************************************/
 /* Pieces                                                   */
